+++ src/sqlfluff/dialects/dialect_mysql.py
@@ -19,6 +19,7 @@ from sqlfluff.core.parser import (
     SymbolSegment,
     Anything,
     Delimited,
+    RegexParser,
 )
 from sqlfluff.core.dialects import load_raw_dialect
 
@@ -129,6 +130,18 @@ mysql_dialect.add(
         ),
         Ref("DatatypeSegment"),
     ),
+    LocalVariableNameSegment=RegexParser(
+        r"`?[a-zA-Z0-9_]*`?",
+        CodeSegment,
+        name="declared_variable",
+        type="variable",
+    ),
+    SessionVariableNameSegment=RegexParser(
+        r"[@][a-zA-Z0-9_]*",
+        CodeSegment,
+        name="declared_variable",
+        type="variable",
+    ),
 )
 
 mysql_dialect.replace(
@@ -136,7 +149,84 @@ mysql_dialect.replace(
     TildeSegment=StringParser(
         "~", SymbolSegment, name="tilde", type="statement_terminator"
     ),
+    ParameterNameSegment=RegexParser(
+        r"`?[A-Za-z0-9_]*`?", CodeSegment, name="parameter", type="parameter"
+    ),
 )
+
+
+@mysql_dialect.segment()
+class DeclareStatement(BaseSegment):
+    """DECLARE statement.
+
+    mysql: https://dev.mysql.com/doc/refman/8.0/en/declare-local-variable.html
+    mysql: https://dev.mysql.com/doc/refman/8.0/en/declare-handler.html
+    mysql: https://dev.mysql.com/doc/refman/8.0/en/declare-condition.html
+    """
+
+    type = "declare_statement"
+
+    match_grammar = OneOf(
+        Sequence(
+            "DECLARE",
+            OneOf("CONTINUE", "EXIT", "UNDO"),
+            "HANDLER",
+            "FOR",
+            OneOf(
+                "SQLEXCEPTION",
+                "SQLWARNING",
+                Sequence("NOT", "FOUND"),
+                Sequence(
+                    "SQLSTATE",
+                    Ref.keyword("VALUE", optional=True),
+                    Ref("QuotedLiteralSegment"),
+                ),
+                OneOf(
+                    Ref("QuotedLiteralSegment"),
+                    Ref("NumericLiteralSegment"),
+                    Ref("NakedIdentifierSegment"),
+                ),
+            ),
+            Sequence(Ref("StatementSegment")),
+        ),
+        Sequence(
+            "DECLARE",
+            Ref("NakedIdentifierSegment"),
+            "CONDITION",
+            "FOR",
+            OneOf(Ref("QuotedLiteralSegment"), Ref("NumericLiteralSegment")),
+        ),
+        Sequence(
+            "DECLARE",
+            Ref("LocalVariableNameSegment"),
+            Ref("DatatypeSegment"),
+            Sequence(
+                Ref.keyword("DEFAULT"),
+                OneOf(
+                    Ref("QuotedLiteralSegment"),
+                    Ref("NumericLiteralSegment"),
+                    Ref("FunctionSegment"),
+                ),
+                optional=True,
+            ),
+        ),
+    )
+
+
+@mysql_dialect.segment(replace=True)
+class StatementSegment(ansi_dialect.get_segment("StatementSegment")):  # type: ignore
+    """Overriding StatementSegment to allow for additional segment parsing."""
+
+    parse_grammar = ansi_dialect.get_segment("StatementSegment").parse_grammar.copy(
+        insert=[
+            Ref("DelimiterStatement"),
+            Ref("CreateProcedureStatementSegment"),
+            Ref("TransactionStatementSegment"),
+            Ref("DeclareStatement"),
+            Ref("SetAssignmentStatementSegment"),
+            Ref("IfExpressionStatement"),
+        ],
+    )
 
 
 @mysql_dialect.segment()
@@ -155,8 +245,7 @@ class CreateProcedureStatementSegment(BaseSegment):
     """
 
     type = "create_procedure_statement"
-    # purposefully left out DEFINER
-    # will be addressed when issue [#1131](https://github.com/sqlfluff/sqlfluff/issues/1131) is addressed
+
     match_grammar = Sequence(
         "CREATE",
         "PROCEDURE",
@@ -175,25 +264,10 @@ class CreateProcedureStatementSegment(BaseSegment):
 
 
 @mysql_dialect.segment(replace=True)
-class StatementSegment(ansi_dialect.get_segment("StatementSegment")):  # type: ignore
-    """Overriding StatementSegment to allow for additional segment parsing."""
-
-    parse_grammar = ansi_dialect.get_segment("StatementSegment").parse_grammar.copy(
-        insert=[
-            Ref("DelimiterStatement"),
-            Ref("CreateProcedureStatementSegment"),
-        ],
-    )
-
-
-@mysql_dialect.segment(replace=True)
 class FunctionDefinitionGrammar(BaseSegment):
     """This is the body of a `CREATE FUNCTION` statement."""
 
-    match_grammar = Sequence(
-        Ref("TransactionStatementSegment"),
-        Ref("StatementSegment"),
-    )
+    match_grammar = Ref("TransactionStatementSegment")
 
 
 @mysql_dialect.segment()
@@ -201,6 +275,7 @@ class CharacteristicStatement(BaseSegment):
     """A Characteristics statement for functions/procedures."""
 
     type = "characteristic_statement"
+
     match_grammar = Sequence(
         OneOf("DETERMINISTIC", Sequence("NOT", "DETERMINISTIC")),
         Sequence("LANGUAGE", "SQL", optional=True),
@@ -229,8 +304,7 @@ class CreateFunctionStatementSegment(BaseSegment):
         "FUNCTION",
         Anything(),
     )
-    # purposefully left out DEFINER
-    # will be addressed when issue [#1131](https://github.com/sqlfluff/sqlfluff/issues/1131) is addressed
+
     parse_grammar = Sequence(
         "CREATE",
         "FUNCTION",
@@ -270,11 +344,103 @@ class DropStatementSegment(BaseSegment):
 class ProcedureParameterListGrammar(BaseSegment):
     """The parameters for a procedure ie. `(in/out/inout name datatype)`."""
 
-    # Function parameter list
     match_grammar = Bracketed(
         Delimited(
             Ref("ProcedureParameterGrammar"),
             delimiter=Ref("CommaSegment"),
             optional=True,
         ),
+    )
+
+
+@mysql_dialect.segment()
+class SetAssignmentStatementSegment(BaseSegment):
+    """A `SET` statement.
+
+    mysql: https://dev.mysql.com/doc/refman/8.0/en/set-variable.html
+    """
+
+    type = "set_statement"
+
+    match_grammar = Sequence(
+        "SET",
+        OneOf(Ref("SessionVariableNameSegment"), Ref("LocalVariableNameSegment")),
+        Ref("EqualsSegment"),
+        AnyNumberOf(
+            Ref("QuotedLiteralSegment"),
+            Ref("DoubleQuotedLiteralSegment"),
+            Ref("SessionVariableNameSegment"),
+            Ref("LocalVariableNameSegment"),
+            Ref("FunctionSegment"),
+        ),
+    )
+
+
+@mysql_dialect.segment(replace=True)
+class TransactionStatementSegment(BaseSegment):
+    """A `COMMIT`, `ROLLBACK` or `TRANSACTION` statement.
+    BEGIN [WORK]
+    COMMIT [WORK] [AND [NO] CHAIN] [[NO] RELEASE]
+    ROLLBACK [WORK] [AND [NO] CHAIN] [[NO] RELEASE]
+    mysql: https://dev.mysql.com/doc/refman/8.0/en/commit.html
+    mysql: https://dev.mysql.com/doc/refman/8.0/en/begin-end.html
+    """
+
+    type = "transaction_statement"
+
+    match_grammar = OneOf(
+        Sequence("START", "TRANSACTION"),
+        Sequence(
+            Sequence(
+                Ref("SingleIdentifierGrammar"), Ref("ColonSegment"), optional=True
+            ),
+            Sequence(
+                "BEGIN",
+                Ref.keyword("WORK", optional=True),
+                Ref("StatementSegment"),
+            ),
+        ),
+        Sequence(
+            "LEAVE",
+            Ref("SingleIdentifierGrammar", optional=True),
+        ),
+        Sequence(
+            "COMMIT",
+            Ref.keyword("WORK", optional=True),
+            Sequence("AND", Ref.keyword("NO", optional=True), "CHAIN", optional=True),
+        ),
+        Sequence(
+            "ROLLBACK",
+            Ref.keyword("WORK", optional=True),
+        ),
+        Sequence(
+            "END",
+            Ref("SingleIdentifierGrammar", optional=True),
+        ),
+    )
+
+
+@mysql_dialect.segment()
+class IfExpressionStatement(BaseSegment):
+    """IF-THEN-ELSE-ELSEIF-END IF statement.
+
+    mysql:https://dev.mysql.com/doc/refman/8.0/en/if.html
+    """
+
+    type = "if_then_statement"
+
+    match_grammar = AnyNumberOf(
+        Sequence(
+            "IF",
+            Ref("ExpressionSegment"),
+            "THEN",
+            Ref("StatementSegment"),
+        ),
+        Sequence(
+            "ELSEIF",
+            Ref("ExpressionSegment"),
+            "THEN",
+            Ref("StatementSegment"),
+        ),
+        Sequence("ELSE", Ref("StatementSegment"), optional=True),
     )