+++ test/fixtures/parser/mysql/begin.sql
@@ -1,0 +1,4 @@
+blocks:BEGIN
+#set @abc = 1;
+select 1;
+END blocks~