+++ test/fixtures/parser/mysql/declare_default_quotedliteral.sql
@@ -1,0 +1,1 @@
+DECLARE abc longtext default 'test';